+++ include/core/commandline.h
@@ -6,6 +6,7 @@
 #include "driver/gpio.h"
 #include "freertos/FreeRTOS.h"
 #include "freertos/task.h"
+#include <stddef.h>
 
 typedef void (*CommandFunction)(int argc, char **argv);
 
@@ -20,6 +21,7 @@ void command_init();
 void register_command(const char *name, CommandFunction function);
 void unregister_command(const char *name);
 CommandFunction find_command(const char *name);
+const char *command_name_at(size_t index);
 void handle_unknown_command(const char *cmd);
 
 extern TaskHandle_t VisualizerHandle;